Prune the extensions that clutter the UI.
Removing the IoT extension because of the way it takes over the UI upon
installing the extension - we recently observed customers attempting to
deploy using the extensions and found they were distracted by the
pop-up. Removed the API Management extension for now since it adds to
the explorer clutter without adding too much value. Removed the MSSQL
extension as it currently has an issue where it adds additional "+"
icons to all other Azure extensions.